pyproject.toml
[tool.poetry]
name = "ask2slide"
version = "0.1.0"
description = ""
authors = ["tang haozhe <tanghaozhe@outlook.com>"]
readme = "README.md"
packages = [
    { include = "llm-server", from = "." }
]

[tool.poetry.dependencies]
colpali_engine = "^0.1.0"
python = "^3.10"
fastapi = "^0.115.12"
uvicorn = "^0.34.2"
minio = "^7.2.15"
pymilvus = "^2.5.8"
torch = ">=2.2.0"
torchvision = ">=0.17.0"
# vllm = {version = "0.4.0", extras = ["triton"], optional = true}  # Temporarily disabled for Mac MPS
transformers = ">=4.39.1"
loguru = "0.7.2"
python-dotenv = "^1.0.0"
accelerate = "^1.7.0"
numpy = "1.26.4"
httpx = "^0.27.0"
python-multipart = "^0.0.6"
pymongo = "^4.6.1"
pillow = "^10.2.0"
tqdm = "^4.66.1"
pydantic-settings = "^2.2.1"

[tool.poetry.extras]
gpu = ["vllm"]

[tool.poetry.scripts]
start-llm = "llm-server.server:main"

[build-system]
requires = [
    "poetry-core>=2.0.0",
    "setuptools>=65.5.1",
    "wheel"
]
build-backend = "poetry.core.masonry.api"
